cs 6680 - #3 - fix a bug in assignment 2
I had to cast a denominator to double. I was mistaken about the default type.

+++ cs_6680_computer_vision/assignment_2/Scaling.m
@@ -44,7 +44,7 @@ assert(input_minimum ~= input_maximum, 'input image has all the same values, res
 %          new maximum - new minimum
 %   k = -------------------------------
 %        input maximum - input minimum
-k = double(range(2) - range(1)) / input_maximum - input_minimum;
+k = double(range(2) - range(1)) / double(input_maximum - input_minimum);
 scaledIm = (inputIm - input_minimum) .* k + range(1);
 transFunc = ([input_minimum:input_maximum] - input_minimum) .* k + range(1);
